Fix missing doxygen dot component in CI.

+++ build_tools/cmake/doxygen.cmake
@@ -2,7 +2,7 @@ MACRO (BUILD_DOC)
     # Build the doc
     # =============
 
-    find_package(Doxygen REQUIRED dot)
+    find_package(Doxygen QUIET COMPONENTS dot)
 
     # Build the C++ documentation.
     configure_file(${CMAKE_SOURCE_DIR}/build_tools/doc_cpp/jiminy.doxyfile
@@ -28,15 +28,19 @@ MACRO (BUILD_DOC)
         VERBATIM
     )
 
-    # Install both documentations.
-    install(
-        CODE "execute_process (COMMAND ${CMAKE_MAKE_PROGRAM} doc)
-              execute_process (COMMAND ${CMAKE_MAKE_PROGRAM} doc_py)
-              file (REMOVE_RECURSE \"${CMAKE_SOURCE_DIR}/docs/cpp\" \"${CMAKE_SOURCE_DIR}/docs/python\")
-              execute_process (COMMAND ${CMAKE_COMMAND} -E copy_directory \"${CMAKE_BINARY_DIR}/doc/html/\" \"${CMAKE_SOURCE_DIR}/docs/cpp\"
-                               COMMAND ${CMAKE_COMMAND} -E copy_directory \"${CMAKE_BINARY_DIR}/doc_py/html/\" \"${CMAKE_SOURCE_DIR}/docs/python\"
-                               )"
-        COMPONENT doxygen
-        EXCLUDE_FROM_ALL
-    )
+    # Install both documentations (if doxygen is available).
+    if(${Doxygen_FOUND})
+        install(
+            CODE "execute_process (COMMAND ${CMAKE_MAKE_PROGRAM} doc)
+                execute_process (COMMAND ${CMAKE_MAKE_PROGRAM} doc_py)
+                file (REMOVE_RECURSE \"${CMAKE_SOURCE_DIR}/docs/cpp\" \"${CMAKE_SOURCE_DIR}/docs/python\")
+                execute_process (COMMAND ${CMAKE_COMMAND} -E copy_directory \"${CMAKE_BINARY_DIR}/doc/html/\" \"${CMAKE_SOURCE_DIR}/docs/cpp\"
+                                COMMAND ${CMAKE_COMMAND} -E copy_directory \"${CMAKE_BINARY_DIR}/doc_py/html/\" \"${CMAKE_SOURCE_DIR}/docs/python\"
+                                )"
+            COMPONENT doxygen
+            EXCLUDE_FROM_ALL
+        )
+    else(${Doxygen_FOUND})
+        message("-- Doxygen with Dot component not found. Documentation generation disabled.")
+    endif(${Doxygen_FOUND})
 ENDMACRO()
